# server/swagger/swagger.json
{"swagger":"2.0","info":{"title":"The appengine example","description":"A simple appengine example","version":""},"host":"localhost:8080","basePath":"/","schemes":["http"],"consumes":["application/json","application/xml","application/gob","application/x-gob"],"produces":["application/json","application/xml","application/gob","application/x-gob"],"paths":{"/memos":{"get":{"tags":["memos"],"summary":"list memos","description":"list","operationId":"memos#list","produces":["application/vnd.goa.error","application/vnd.memo+json; type=collection"],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/MemoCollection"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/error"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]},"post":{"tags":["memos"],"summary":"create memos","description":"create","operationId":"memos#create","produces":["application/vnd.goa.error","application/vnd.memo+json"],"parameters":[{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/MemoPayload"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/Memo"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/error"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}},"/memos/{id}":{"get":{"tags":["memos"],"summary":"show memos","description":"show","operationId":"memos#show","produces":["application/vnd.goa.error","application/vnd.memo+json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Memo"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/error"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]},"put":{"tags":["memos"],"summary":"update memos","description":"update","operationId":"memos#update","produces":["application/vnd.goa.error","application/vnd.memo+json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string"},{"name":"payload","in":"body","required":true,"schema":{"$ref":"#/definitions/MemoPayload"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Memo"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/error"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]},"delete":{"tags":["memos"],"summary":"delete memos","description":"delete","operationId":"memos#delete","produces":["application/vnd.goa.error","application/vnd.memo+json"],"parameters":[{"name":"id","in":"path","required":true,"type":"string"},{"name":"name","in":"query","required":false,"type":"string"},{"name":"org_id","in":"query","description":"Organization ID","required":false,"type":"string"}],"responses":{"204":{"description":"No Content","schema":{"$ref":"#/definitions/Memo"}},"400":{"description":"Bad Request","schema":{"$ref":"#/definitions/error"}},"401":{"description":"Unauthorized","schema":{"$ref":"#/definitions/error"}},"404":{"description":"Not Found","schema":{"$ref":"#/definitions/error"}},"409":{"description":"Conflict","schema":{"$ref":"#/definitions/error"}},"500":{"description":"Internal Server Error","schema":{"$ref":"#/definitions/error"}}},"schemes":["http"]}}},"definitions":{"Memo":{"title":"Mediatype identifier: application/vnd.memo+json; view=default","type":"object","properties":{"content":{"type":"string","description":"Content of memo","example":"Hic tenetur ut."},"created_at":{"type":"string","description":"Time when memo is created","example":"2008-05-09T17:01:53Z","format":"date-time"},"id":{"type":"string","example":"Ratione ipsam qui blanditiis numquam consequatur."},"shared":{"type":"boolean","description":"Shared to public","example":true},"updated_at":{"type":"string","description":"Time when memo is updated","example":"2007-03-31T16:05:01Z","format":"date-time"}},"description":"Memo (default view)","example":{"content":"Hic tenetur ut.","created_at":"2008-05-09T17:01:53Z","id":"Ratione ipsam qui blanditiis numquam consequatur.","shared":true,"updated_at":"2007-03-31T16:05:01Z"},"required":["id","content","shared","created_at","updated_at"]},"MemoCollection":{"title":"Mediatype identifier: application/vnd.memo+json; type=collection; view=default","type":"array","items":{"$ref":"#/definitions/Memo"},"description":"MemoCollection is the media type for an array of Memo (default view)","example":[{"content":"Hic tenetur ut.","created_at":"2008-05-09T17:01:53Z","id":"Ratione ipsam qui blanditiis numquam consequatur.","shared":true,"updated_at":"2007-03-31T16:05:01Z"},{"content":"Hic tenetur ut.","created_at":"2008-05-09T17:01:53Z","id":"Ratione ipsam qui blanditiis numquam consequatur.","shared":true,"updated_at":"2007-03-31T16:05:01Z"}]},"MemoPayload":{"title":"MemoPayload","type":"object","properties":{"content":{"type":"string","description":"Content of memo","example":"Sed officia."},"shared":{"type":"boolean","description":"Shared to public","example":true}},"example":{"content":"Sed officia.","shared":true},"required":["content"]},"error":{"title":"Mediatype identifier: application/vnd.goa.error; view=default","type":"object","properties":{"code":{"type":"string","description":"an application-specific error code, expressed as a string value.","example":"invalid_value"},"detail":{"type":"string","description":"a human-readable explanation specific to this occurrence of the problem.","example":"Value of ID must be an integer"},"id":{"type":"string","description":"a unique identifier for this particular occurrence of the problem.","example":"3F1FKVRR"},"meta":{"type":"object","description":"a meta object containing non-standard meta-information about the error.","example":{"timestamp":1458609066},"additionalProperties":true},"status":{"type":"string","description":"the HTTP status code applicable to this problem, expressed as a string value.","example":"400"}},"description":"Error response media type (default view)","example":{"code":"invalid_value","detail":"Value of ID must be an integer","id":"3F1FKVRR","meta":{"timestamp":1458609066},"status":"400"}}}}